Complete initialisation
Since we'll rely on a custom User model for our app, we definitely need
to include this in initial setup. This must be done before initial
`manage makemigrations` otherwise it will result in inconsistent
migration history.

Start basic models structure and admin customisation. Far from being
complete but it's a start for the rest of the project.

Start centralisation of dependencies inside `pyproject.toml` instead of
mixed sources (requirements.in + pyproject.toml). Goal being to totally
remove `requirements.in`.