Flatten migration history & continue implem
Since we have no data / no app available in the wild I took the liberty
to flatten migrations to avoid complex migration path and simply
recreate the DB. Might do that again if big model mistakes are detected.

Add missing dependencies (environs) and update lock + requirements.txt.

Homogenize commands in justfile to rely on `docker-compose` instead of
`docker compose`. Not strictly needed but it's the newest way of dealing
with docker-compose so I think it's more durable.