Add seed command and fix nullable model values
Create a `seed` management command to quickly start testing project
features. Both for me and admin users to allow rapid setup.

Since the chosen design rely on the fact we will populate movie and
authors with TMDB we need to support null values for the majority of the
fields. Reset migrations to avoid complex migration path on non-existing
database.